+++ src/egon/data/datasets/sanity_checks.py
@@ -90,8 +90,7 @@ def etrago_eGon2035_electricity():
     )
 
     carriers_electricity = [
-        "other_non_renewable",
-        "other_renewable",
+        "others",
         "reservoir",
         "run_of_river",
         "oil",
@@ -325,8 +324,8 @@ def etrago_eGon2035_heat():
     None
     """
 
-    # Check input and output values for the carriers "other_non_renewable",
-    # "other_renewable", "reservoir", "run_of_river" and "oil"
+    # Check input and output values for the carriers "others",
+    # "reservoir", "run_of_river" and "oil"
 
     scn = "eGon2035"